// src/selectors/funnel/models/IFunnelTableRow.ts
export default interface IFunnelTableRow {
  readonly funnelGenID: number;
  readonly funnelID: string;
  readonly salesID: number;
  readonly salesName: string;
  readonly presalesName: string;
  readonly productManager: string;
  readonly customerName: string;
  readonly projectName: string;
  readonly totalSellingPrice: number;
  readonly totalOrderingItem: number;
  readonly totalSellingItem: number;
  readonly gpmPctg: number;
  readonly gpmAmount: number;
  readonly dealCloseDate: string;
  readonly createDate: string;
  readonly funnelStatus: string;
  readonly customerGenID: number;
  readonly dept: string;
  readonly commercialWorkflowStatus: string;
  readonly serviceWorkflowStatus: string;
  readonly flagSA: number;
  readonly flagOpen: string;
  readonly soidc: string;
  readonly currency: string;
  readonly rate: number;
  readonly reassignFlag: number;
  readonly soParent: string;
  readonly saDate: string;
  readonly saNumber: string;
  readonly stepName: string;
  readonly flagManual: string;
}